+++ library/terraform/civo/firewall.tf
@@ -1,0 +1,5 @@
+resource "civo_firewall" "your_firewall" {
+  name       = "your-firewall-name"
+  network_id = civo_network.your_network.id
+  create_default_rules = true
+}